feat: Further customization.

+++ scripts/install-configuration.ps1
@@ -30,50 +30,56 @@ if (-not ([Security.Principal.WindowsPrincipal][Security.Principal.WindowsIdenti
 # ----------------------------------------
 
 if (((Get-ItemProperty -Path $installationKey -Name "$($ShellType)Installed")."$($ShellType)Installed") -eq 1) {
-    Remove-Item -Path "HKCU:\Software\Azrael\$($ShellType)" -Force -Recurse -ErrorAction SilentlyContinue
-    Remove-ItemProperty -Path $installationKey -Name "$($ShellType)Installed" -Force -ErrorAction SilentlyContinue
-    Remove-Item -Path "$($PROFILE | Split-Path)\*" -Force -Recurse -ErrorAction SilentlyContinue
-    Write-Host "Shell configuration is already installed. Exiting... re-run the script to reinstall."
-    exit
+    Write-Host "Shell configuration is already installed. Do you want to reinstall it? (y/n)"
+    $input = Read-Host
+    if ($input -eq "y") {
+        Write-Host "Reinstalling..."
+        Remove-Item -Path "HKCU:\Software\Azrael\$($ShellType)" -Force -Recurse -ErrorAction SilentlyContinue
+        Remove-ItemProperty -Path $installationKey -Name "$($ShellType)Installed" -Force -ErrorAction SilentlyContinue
+        Remove-Item -Path "$($PROFILE | Split-Path)\*" -Force -Recurse -ErrorAction SilentlyContinue
+    }
+    else {
+        Write-Host "Exiting..."
+        exit
+    }
 } 
-else {
-    Write-Host "Shell configuration is not installed. Installing..."
-    if (-not (Test-Path $installationKey)) {
-        New-Item -Path $installationKey -Force
-    }
-
-    if (-not (Test-Path $PROFILE)) {
-        New-Item -Path $PROFILE -Force
-    }
-
-    Write-Host "Microsoft Windows [Version $($KernelVersion)]" # Fun. I like it.
-    Write-Host "(c) Microsoft Corporation. All rights reserved.`n"
-    Write-Host "Copyright (c) 2023-2024 Azrael"
-    Write-Host "https://github.com/luke-beep/shell-config/`n"
-    
-    Write-Host "This script covers the following shells: PowerShell, PowerShell Core, Windows Terminal & Command Prompt/Clink."
-        
-    Write-Host "Installing the necessary packages..."
-    Start-Process powershell -Verb runAs -ArgumentList "-NoProfile -Command `"& { Invoke-Expression (New-Object System.Net.WebClient).DownloadString('https://raw.githubusercontent.com/luke-beep/shell-config/main/scripts/packages/install-packages.ps1') }`""
-    Write-Host "Packages installed."
-
-    Write-Host "Applying Windows Terminal settings..."
-    Invoke-WebRequest -Uri https://github.com/luke-beep/shell-config/raw/main/configs/terminal/settings.json -OutFile $TerminalSettingsPath
-    Write-Host "Warning: These settings may cause issues with Windows Terminal. Please check the settings.json file in $TerminalSettingsPath. If you encounter any issues, please open an issue on GitHub."
-
-    Write-Host "Installing the oh-my-posh theme for Clink..."
-    if (-not (Test-Path $OmpConfig)) {
-        New-Item -Path $OmpConfig -Force 
-        Invoke-WebRequest -Uri https://raw.githubusercontent.com/luke-beep/shell-config/main/configs/omp/theme.json -OutFile $OmpConfig
-    }
-    Invoke-WebRequest -Uri https://github.com/luke-beep/shell-config/raw/main/configs/clink/oh-my-posh.lua -OutFile $ClinkThemePath
-
-    Write-Host "Installing the $ShellType configuration..."
-    Invoke-WebRequest -Uri https://github.com/luke-beep/shell-config/raw/main/configs/pwsh/Microsoft.PowerShell_profile.ps1 -OutFile $PROFILE
-
-    Write-Host "Installation complete. Please restart your shell."
-    New-ItemProperty -Path $installationKey -Name "$($ShellType)Installed" -Value 1 -PropertyType DWORD -Force
+Write-Host "Shell configuration is not installed. Installing..."
+if (-not (Test-Path $installationKey)) {
+    New-Item -Path $installationKey -Force
 }
+
+if (-not (Test-Path $PROFILE)) {
+    New-Item -Path $PROFILE -Force
+}
+
+Write-Host "Microsoft Windows [Version $($KernelVersion)]" # Fun. I like it.
+Write-Host "(c) Microsoft Corporation. All rights reserved.`n"
+Write-Host "Copyright (c) 2023-2024 Azrael"
+Write-Host "https://github.com/luke-beep/shell-config/`n"
+    
+Write-Host "This script covers the following shells: PowerShell, PowerShell Core, Windows Terminal & Command Prompt/Clink."
+        
+Write-Host "Installing the necessary packages..."
+Start-Process powershell -Verb runAs -ArgumentList "-NoProfile -Command `"& { Invoke-Expression (New-Object System.Net.WebClient).DownloadString('https://raw.githubusercontent.com/luke-beep/shell-config/main/scripts/packages/install-packages.ps1') }`"" -Wait
+Write-Host "Packages installed."
+
+Write-Host "Applying Windows Terminal settings..."
+Invoke-WebRequest -Uri https://github.com/luke-beep/shell-config/raw/main/configs/terminal/settings.json -OutFile $TerminalSettingsPath -ErrorAction SilentlyContinue
+Write-Host "Warning: These settings may cause issues with Windows Terminal. Please check the settings.json file in $TerminalSettingsPath. If you encounter any issues, please open an issue on GitHub."
+
+Write-Host "Installing the oh-my-posh theme for Clink..."
+if (-not (Test-Path $OmpConfig)) {
+    New-Item -Path $OmpConfig -Force 
+    Invoke-WebRequest -Uri https://raw.githubusercontent.com/luke-beep/shell-config/main/configs/omp/theme.json -OutFile $OmpConfig -ErrorAction SilentlyContinue
+}
+Invoke-WebRequest -Uri https://github.com/luke-beep/shell-config/raw/main/configs/clink/oh-my-posh.lua -OutFile $ClinkThemePath -ErrorAction SilentlyContinue
+
+Write-Host "Installing the $ShellType configuration..."
+Invoke-WebRequest -Uri https://github.com/luke-beep/shell-config/raw/main/configs/pwsh/Microsoft.PowerShell_profile.ps1 -OutFile $PROFILE -ErrorAction SilentlyContinue
+
+Write-Host "Installation complete. Please restart your shell."
+New-ItemProperty -Path $installationKey -Name "$($ShellType)Installed" -Value 1 -PropertyType DWORD -Force
+
 
 # ----------------------------------------
 # End
